cr changes
Former-commit-id: 7ea9da00bea7a5a31a0e0750bd2c2b4a24db53e0

+++ cmd/lakefs/cmd/migrate.go
@@ -93,6 +93,6 @@ func init() {
 	migrateCmd.AddCommand(gotoCmd)
 	_ = gotoCmd.Flags().String("schema", "", "schema name")
 	_ = gotoCmd.MarkFlagRequired("schema")
-	_ = gotoCmd.Flags().Uint("version", uint(0), "version number")
+	_ = gotoCmd.Flags().Uint("version", 0, "version number")
 	_ = gotoCmd.MarkFlagRequired("version")
 }
